requirements.txt
# ChatGPT WhatsApp & Messenger Webhook Requirements

# Web Framework
flask==3.0.0
asgiref==3.7.2
uvicorn==0.24.0

# HTTP Client
aiohttp==3.9.1

# OpenAI API
openai==0.28.1

# Environment Variables
python-dotenv==1.0.0

# WebSocket Support for Real-time Web Interface
# Core Flask and Web Framework
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flask-socketio==5.3.6
python-dotenv==1.0.1

# Database and ORM
PyMySQL==1.1.1
cryptography==42.0.8
Werkzeug==3.0.3

# AI and ML Dependencies
openai==0.28
scikit-learn==1.5.1
spacy==3.7.5

# HTTP and Async
aiohttp==3.10.5
requests==2.32.3

# Server and ASGI
asgiref==3.8.1
uvicorn[standard]==0.30.6
eventlet==0.36.1

# Data Processing
numpy==1.26.4
pandas==2.2.2

# Text Processing and NLP
unicodedata2==15.1.0
difflib2==0.1

# Utilities
pathlib2==2.3.7
typing-extensions==4.12.2

# Development and Testing
python-json-logger==2.0.7
eventlet==0.33.3

# Async Support
asyncio